+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/CoreModuleConfig.java
@@ -231,6 +231,20 @@ public class CoreModuleConfig extends ModuleConfig {
     @Getter
     private boolean enableHierarchy = true;
 
+    /**
+     * The int value of the max heap memory usage percent.
+     * The default value is 85%.
+     */
+    @Getter
+    private long maxHeapMemoryUsagePercent = 85;
+
+    /**
+     * The long value of the max direct memory usage.
+     * The default max value is -1, representing no limit.
+     */
+    @Getter
+    private long maxDirectMemoryUsage = -1;
+
     public CoreModuleConfig() {
         this.downsampling = new ArrayList<>();
     }