+++ oap-server/server-telemetry/telemetry-api/src/main/java/org/apache/skywalking/oap/server/telemetry/none/MetricsCollectorNoop.java
@@ -30,4 +30,19 @@ public class MetricsCollectorNoop implements MetricsCollector {
     public Iterable<MetricFamily> collect() {
         return Collections.emptyList();
     }
+
+    @Override
+    public long heapMemoryUsage() {
+        return 0;
+    }
+
+    @Override
+    public long heapMemoryMax() {
+        return 0;
+    }
+
+    @Override
+    public long directMemoryUsage() {
+        return 0;
+    }
 }